\name{BOINETC-package}
\alias{BOINETC-package}
\alias{BOINETC}
\title{BOIN-ETC Simulation Utilities}
\description{Tools for simulating and summarizing BOIN-ETC dose-combination designs assembled from the provided R scripts.}
\details{The package includes core functions for subtrial dose selection, BOIN-ETC1/2/3 operating-characteristic simulations, ODC summaries, and convenience workflow helpers.}
\keyword{package}
